==> 1. Preprocess.ipynb <==
<font color='red'><h1> Workbook for creating validation set and data preprocessing </h1></font>

This workbook is to
1. Creating validation set
2. Drop unnecessary columns
3. Fill missing values
4. Standard scaling

## Import libraries

import os
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler

csv_path = os.getcwd() + '/data/reserved/train.csv'
df = pd.read_csv(csv_path)

## Drop columns

df.drop(df.columns[0],axis=1,inplace=True)

cols = df.columns[56:63]
df.drop(columns=cols,inplace=True)

## Fill missing values

for col in df.columns:
    df[col] = df[col].fillna(df[col].median())

## Split dataset

X = df.drop('emission',axis=1)
y = df['emission']

X_train, X_test, y_train, y_test = train_test_split(X,y,test_size=0.2)

## Standardisation

scaler = StandardScaler()

X_train = scaler.fit_transform(X_train)
X_test = scaler.fit_transform(X_test)

X_train = pd.DataFrame(X_train, columns=df.drop('emission',axis=1).columns)
X_test = pd.DataFrame(X_test, columns=df.drop('emission',axis=1).columns)

X_train['emission'] = y_train.values
X_test['emission'] = y_test.values

X_train.to_csv("data/train.csv",index=False)

X_test.to_csv("data/val.csv",index=False)
