==> 2.1 Row training NN.ipynb <==
<font color='red'><h1> Training Neural Network on row data </h1></font>

This workbook is to train neural network on row data. By row data, I mean data without feature engineering. This workbook is to
1. Find perfect hyperparameters using keras_tuner
2. Train NN on train dataset

## Import libraries

import os,shutil
import numpy as np
import pandas as pd
import tensorflow as tf
import keras_tuner
from tensorflow import keras
from keras import Sequential
from keras.callbacks import EarlyStopping
from keras.layers import Dense, Dropout, BatchNormalization
from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier, GradientBoostingClassifier

csv_path = os.getcwd() + '/data/'
train_df = pd.read_csv(csv_path + 'train.csv')
val_df = pd.read_csv(csv_path + 'val.csv')

X_train = train_df.drop('emission',axis=1)
X_val = val_df.drop('emission',axis=1)
y_train = train_df['emission']
y_val = val_df['emission']

## Tune the model

neurons1 = np.arange(200,301,8).tolist()
neurons2 = np.arange(100,201,8).tolist()
neurons3 = np.arange(50,101,5).tolist()
neurons4 = np.arange(10,51,5).tolist()

new_model = keras.Sequential()
new_model.add(keras.Input(67,))

def build_model(hp):
    model = keras.Sequential()
    
    if len(new_model.layers) > 0:
        model.add(new_model)
    
    hp_units = hp.Choice('units', neurons4)
    model.add(Dense(hp_units,activation='relu'))
    
    model.add(Dense(1,activation='linear'))
    
    model.compile(loss='mse',optimizer='adam',metrics=['mae'])
    return model

tuner = keras_tuner.RandomSearch(build_model,objective='val_loss',max_trials=5)

tuner.search(X_train, y_train, epochs=3, validation_data=(X_val, y_val))

best_model = tuner.get_best_models()[0]

new_model = tf.keras.Model(inputs=best_model.input, outputs=best_model.layers[-2].output)
for layer in new_model.layers:
    layer.trainable = False

shutil.rmtree(os.getcwd() + '/untitled_project')

## Tune NN

callback = EarlyStopping(monitor='val_loss',min_delta=0,patience=10,mode='auto')

model = keras.Sequential()

model.add(Dense(375,activation='relu'))
model.add(Dense(300,activation='relu'))
model.add(Dense(200,activation='relu'))

model.add(Dense(150,activation='relu'))
model.add(Dropout(0.2))
model.add(Dense(100,activation='relu'))
model.add(Dropout(0.25))

model.add(Dense(75,activation='relu'))
model.add(Dropout(0.2))
model.add(Dense(50,activation='relu'))
model.add(Dropout(0.2))

model.add(Dense(37,activation='relu'))
model.add(Dropout(0.2))
model.add(Dense(32,activation='relu'))
model.add(Dropout(0.2))

model.add(Dense(16,activation='relu'))
model.add(Dropout(0.2))
model.add(Dense(8,activation='relu'))
model.add(Dropout(0.2))


model.add(Dense(4,activation='relu'))
model.add(Dropout(0.25))

model.add(Dense(2,activation='relu'))
model.add(Dense(1,activation='linear'))

model.compile(loss='mse',optimizer='rmsprop',metrics=['mae'])

model.fit(X_train,y_train,epochs=100,batch_size=32,validation_data=(X_val,y_val),callbacks=callback)
