==> 2.2 Row training ML.ipynb <==
<font color='red'><h1> Workbook to train traditional ML models on row data </h1></font>

This workbook is to train ML models perticularly bagging and boosting tree based models on row data, by row data I mean data without feature engineering. This workbook is to
1. Tune and train random forest regressor
2. Tune and train gradient boosting regressor
3. Tune and train ada boost regressor
4. Tune and train xgboost regressor

## Import libraries

import os
import pandas as pd
import numpy as np
import xgboost as xgb
from sklearn.metrics import mean_absolute_error
from sklearn.ensemble import RandomForestRegressor, AdaBoostRegressor, GradientBoostingRegressor

## Process data

csv_path = os.getcwd() + '/data/'
train_df = pd.read_csv(csv_path + 'train.csv')
val_df = pd.read_csv(csv_path + 'val.csv')

X_train = train_df.drop('emission',axis=1)
X_val = val_df.drop('emission',axis=1)
y_train = train_df['emission']
y_val = val_df['emission']

## Random Forest

rf = RandomForestRegressor(n_estimators=150)
rf.fit(X_train,y_train)

y_pred = rf.predict(X_val)
mean_absolute_error(y_val,y_pred)

## Gradient Boost

gbr = GradientBoostingRegressor(n_estimators=100,max_depth=15)
gbr.fit(X_train,y_train)

y_pred = gbr.predict(X_val)
mean_absolute_error(y_val,y_pred)

## AdaBoost

adb = AdaBoostRegressor(n_estimators=220)
adb.fit(X_train,y_train)

y_pred = adb.predict(X_val)
mean_absolute_error(y_val,y_pred)

## XGBoost

xgb = xgb.XGBRegressor(objective='reg:squarederror',n_estimators=107,learning_rate=0.1,max_depth=13,random_state=42)
xgb.fit(X_train, y_train)

y_pred = xgb.predict(X_val)
mean_absolute_error(y_val,y_pred)
