==> 3.2 BorutaPy.ipynb <==
<font color='red'><h1> Workbook for feature ranking based on BorutaPy library </h1></font>

## Import libraries

import numpy as np
import pandas as pd
import os
from sklearn.ensemble import RandomForestRegressor
from boruta import BorutaPy

csv_path = os.getcwd() + "/data/"

train_df = pd.read_csv(csv_path + "train.csv")
val_df = pd.read_csv(csv_path + "val.csv")

X_train = train_df.drop('emission',axis=1)
y_train = train_df['emission']
X_val = val_df.drop('emission',axis=1)
y_val = val_df['emission']

## BorutaPy

# define random forest classifier
forest = RandomForestRegressor(max_depth=30,n_jobs=-1)

# define Boruta feature selection method
feat_selector = BorutaPy(forest, n_estimators='auto', verbose=2, random_state=1,max_iter=50)

# find all relevant features
feat_selector.fit(X_train.values,y_train.values)

ranks = feat_selector.ranking_
params = train_df.columns

sorted_pairs = sorted(zip(ranks, params))
sorted_first_list, sorted_second_list = zip(*sorted_pairs)

import csv
with open(csv_path+'ranked_features.csv', 'w', newline='') as csvfile:
    writer = csv.writer(csvfile)

    # Write the header row if needed
    writer.writerow(['Rank', 'Feature'])

    # Iterate over the arrays and write each element as a row
    for i in range(len(sorted_first_list)):
        writer.writerow([sorted_first_list[i], sorted_second_list[i]])
