==> 6.2 F3 training ML.ipynb <==
<font color='red'><h1> Workbook to train bagging and boosting ML models on BorutaPy ranked features. </h1></font>

This workbook is to train bagging and boosting tree based ML models from different set of ranked features. The features are ranked by their importance by BorutaPy library where rank 1 means the best one.

## Import libraries

import os
import pandas as pd
import numpy as np
import xgboost as xgb
from sklearn.metrics import mean_absolute_error
from sklearn.ensemble import RandomForestRegressor, AdaBoostRegressor, GradientBoostingRegressor

## Process data

csv_path = os.getcwd() + '/data/'
train_df = pd.read_csv(csv_path + 'train.csv')
val_df = pd.read_csv(csv_path + 'val.csv')
feature_df = pd.read_csv(csv_path + 'ranked_features.csv')

X_train = train_df.drop('emission',axis=1)
X_val = val_df.drop('emission',axis=1)
y_train = train_df['emission']
y_val = val_df['emission']

## Train Model

def train_model(X_train,y_train,X_val,y_val):
    
    xgb_model = xgb.XGBRegressor(objective='reg:squarederror',n_estimators=100,learning_rate=0.1,max_depth=13,random_state=42)
    xgb_model.fit(X_train, y_train)
    
    y_pred = xgb_model.predict(X_val)
    print(mean_absolute_error(y_val,y_pred))

## Rank 1 only

cols = feature_df[feature_df['Rank'] == 1]['Feature'].values

X_train_new = X_train[cols]
X_val_new = X_val[cols]

train_model(X_train_new,y_train,X_val_new,y_val)

## Rank upto 10

cols = feature_df[feature_df['Rank'] <= 10]['Feature'].values

X_train_new = X_train[cols]
X_val_new = X_val[cols]

train_model(X_train_new,y_train,X_val_new,y_val)

## Rank upto 20

cols = feature_df[feature_df['Rank'] <= 20]['Feature'].values

X_train_new = X_train[cols]
X_val_new = X_val[cols]

train_model(X_train_new,y_train,X_val_new,y_val)
