==> 7. Deploy Model.ipynb <==
<font color='red'><h1> Workbook to deploy the best strategy </h1></font>

This workbook deploys the best strategy, which includes the top 20 features ranked by BorutaPy library and XGBoost regressor with hyperparameter tuning.

## Import libraries

import numpy as np
import pandas as pd
import os
import xgboost as xgb
from sklearn.preprocessing import StandardScaler

csv_path = os.getcwd() + "/data/reserved/"

train_df = pd.read_csv(csv_path + "train.csv")
test_df = pd.read_csv(csv_path + "test.csv")
feature_df = pd.read_csv(os.getcwd() + '/data/ranked_features.csv')

## Preprocessing

cols = feature_df[feature_df['Rank'] <= 10]['Feature'].values

test_df = test_df[cols]
train_df = train_df[np.append(cols,['emission'])]

for col in test_df.columns:
    train_df[col] = train_df[col].fillna(train_df[col].median())
    test_df[col] = test_df[col].fillna(test_df[col].median())

X_train = train_df.drop('emission',axis=1)
y_train = train_df['emission']

scaler = StandardScaler()

X_train = scaler.fit_transform(X_train)
test_df = scaler.fit_transform(test_df)

X_train = pd.DataFrame(X_train, columns=train_df.drop('emission',axis=1).columns)
test_df = pd.DataFrame(test_df, columns=train_df.drop('emission',axis=1).columns)

## Deploy the model

xgb_model = xgb.XGBRegressor(objective='reg:squarederror',n_estimators=100,learning_rate=0.1,max_depth=13,random_state=42)
xgb_model.fit(X_train, y_train)

y_pred = xgb_model.predict(test_df)

## Make submission

df = pd.read_csv(csv_path + 'test.csv')
df = df[df.columns[0:2]]

df['emission'] = y_pred
df.drop('latitude',axis=1,inplace=True)

df.to_csv(csv_path + "submission.csv",index=False)
